Renamed DSAVerificator.cs -> DSAChecker.cs Refactored some enums
